+++ src/src/devtools.ts
@@ -1,6 +1,6 @@
 chrome.devtools.panels.create(
     'http helper',
-    'dist/icons/icon16.png',
-    'dist/panel.html',
+    'icons/icon16.png',
+    'panel.html',
     () => console.log('[http helper] Panel created')
 )